Add SageAttention backends, torch.compile, and inference performance fixes
Introduce a centralized attention dispatch system (attention_dispatch.py)
that supports three backends: PyTorch SDPA (default), SageAttention v2
(Ampere+, ~2x faster via INT8 Q/K quantization), and SageAttention v3
(Blackwell only, ~3x faster via FP4 microscaling quantization). Both sage
backends are drop-in replacements for SDPA with negligible quality loss
at inference. The dispatcher automatically falls back to SDPA when
attention masks are present since sage backends don't reliably support
them — this preserves correctness for multi-view global attention blocks.
Sage kernel calls are wrapped with @torch.compiler.disable to prevent
dynamo tracing into opaque CUDA extensions.

Add per-block torch.compile on individual ViT transformer blocks rather
than whole-model compilation. Individual blocks are clean subgraphs
(norm → attn → residual → norm → mlp → residual) that compile without
graph breaks when using SDPA, enabling kernel fusion across LayerNorm,
linear projections, and residual additions. With sage backends, each
block has one graph break at the attention dispatch but norm/mlp/projection
ops still get fused. Uses dynamic=False since input sizes are fixed after
patch-size alignment, allowing shape-specialized kernels without guard
overhead.

Both options are exposed as user-facing parameters on the model loader
node (attention_mode and torch_compile).

Additional performance fixes:

- Remove blanket autocast(enabled=False) around the entire DPT head in
  da3.py. This was forcing all Conv2d/ConvTranspose2d/Linear layers in
  the DPT fusion pyramid to run in float32 even when bf16/fp16 precision
  was selected. The only ops that actually need float32 are exp() calls
  which overflow in fp16 (max ~65504, exp overflows at x > 11.09). Add
  targeted .float()/.to(x.dtype) casts only around exp/expm1 in
  dpt.py, dualdpt.py, and head_utils.py — these tensors are small at
  that point ([B, H, W, 1-2]) so the cast cost is negligible, while the
  conv layers throughout the head now benefit from half-precision compute.

- Remove nn.quantized.FloatFunctional from ResidualConvUnit and
  FeatureFusionBlock in dpt.py. FloatFunctional is a quantization-aware
  wrapper that causes torch.compile graph breaks (~24 per DPT head
  forward pass). Replace with plain tensor addition since these modules
  have no quantization applied. FloatFunctional has no learnable
  parameters so checkpoint compatibility is maintained via strict=False
  loading.

- Cache interpolated position embeddings in vision_transformer.py keyed
  by (patch_grid_w, patch_grid_h, dtype). The position embeddings are a
  fixed function of model weights and input spatial dimensions — they
  don't depend on pixel content. Without the cache, every forward pass
  recomputes a bicubic interpolation, float32 upcast/downcast, reshape,
  permute, and concatenation. In typical ComfyUI workflows where images
  share the same resolution, this computation is redundant after the
  first call.

- Use expand() instead of repeat() for attention mask broadcasting
  across heads in dinov2/layers/attention.py. repeat() allocates a new
  tensor and physically copies data num_heads times; expand() returns a
  strided view of the same memory with zero copy and zero allocation.
  The result is identical since SDPA only reads the mask.

+++ nodes/depth_anything_v3/model/dualdpt.py
@@ -347,11 +347,11 @@ class DualDPT(nn.Module):
         """
         act = activation.lower() if isinstance(activation, str) else activation
         if act == "exp":
-            return torch.exp(x)
+            return torch.exp(x.float()).to(x.dtype)
         if act == "expm1":
-            return torch.expm1(x)
+            return torch.expm1(x.float()).to(x.dtype)
         if act == "expp1":
-            return torch.exp(x) + 1
+            return (torch.exp(x.float()) + 1).to(x.dtype)
         if act == "relu":
             return torch.relu(x)
         if act == "sigmoid":
